reporting_platform/requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
Flask_SQLAlchemy==3.1.1
requests
scikit-learn
nltk
python-dotenv
pillow
schedule
gunicorn
